Refactor glucose dashboard: implement date range selector, enhance chart generation, and modularize chart functions for improved readability and maintainability.
