+++ glucose_dashboard/charts/__init__.py
@@ -1,0 +1,7 @@
+# charts/__init__.py
+
+print("[charts] __init__.py loaded")
+
+from .time_series import generate_glucose_time_chart
+from .heatmap import generate_glucose_heatmap
+from .daily_avg import generate_daily_average_chart
